% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmoswing.R
\docType{package}
\name{atmoswing-package}
\alias{atmoswing}
\alias{atmoswing-package}
\title{AtmoSwing R toolbox.}
\description{
Postprocessing of the AtmoSwing results in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoswing/R-toolbox}
  \item Report bugs at \url{https://github.com/atmoswing/R-toolbox/issues}
}

}
\author{
\strong{Maintainer}: Pascal Horton \email{pascal.horton@giub.unibe.ch}

Other contributors:
\itemize{
  \item Kathrin Burkart [contributor]
}

}
