% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmoswingRToolbox.R
\docType{package}
\name{atmoswingRToolbox-package}
\alias{atmoswingRToolbox}
\alias{atmoswingRToolbox-package}
\title{AtmoSwing R toolbox.}
\description{
Postprocessing of the AtmoSwing results in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoswing/R-toolbox}
  \item Report bugs at \url{https://github.com/atmoswing/R-toolbox/issues}
}

}
\author{
\strong{Maintainer}: Pascal Horton \email{pascal.horton@giub.unibe.ch}

Authors:
\itemize{
  \item Kathrin B.
}

}
