% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colMax}
\alias{colMax}
\title{Get the column max value.}
\usage{
colMax(data)
}
\value{
The maximum value found in the column
}
\description{
Get the column max value.
}
\examples{
\dontrun{
max <- atmoswing::colMax(data)
}

}
