% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colMin}
\alias{colMin}
\title{Get the column min value.}
\usage{
colMin(data)
}
\value{
The minimum value found in the column
}
\description{
Get the column min value.
}
\examples{
\dontrun{
min <- atmoswing::colMin(data)
}

}
