% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutputs.R
\name{createStationsDataframe}
\alias{createStationsDataframe}
\title{Build a dataframe to store results.}
\usage{
createStationsDataframe()
}
\value{
The empty dataframe.
}
\description{
Build an empty dataframe to store the stations properties.
}
\examples{
\dontrun{
stations <- atmoswing::createStationsDataframe()
}

}
