% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crps}
\alias{crps}
\title{Continuous ranked probability score}
\usage{
crps(x, x0, a = 0.44, b = 0.12)
}
\arguments{
\item{x}{vector of length n (number of members / analogues)}

\item{x0}{observed value}

\item{a, b}{constants that have a law-dependent optimum from which the samples are derived}
}
\value{
Value of the CRPS score
}
\description{
Process the CRPS score for a given probability prediction compared to an observed value. 
The non-exceedence freequency is processed as follows: \code{F(x) = (r-a)/(n+b)}
}
\examples{
obs <- 25.8
precip <- c(35.3, 31.5, 14.9, 30.0, 3.0, 4.2, 15.1, 1.9, 1.2, 2.1)
result <- atmoswingRToolbox::crps(precip, obs, a=0.375, b=0.25)

}
