% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crpsContribClasses}
\alias{crpsContribClasses}
\title{Get the contribution to the CRPS per predictand classes.}
\usage{
crpsContribClasses(crps.vect, obs.vect)
}
\arguments{
\item{crps.vect}{A vector of the CRPS value for every target day.}

\item{obs.vect}{A vector of the observed value (same length as crps.vect).}
}
\value{
Contribution to the CRPS (%).
}
\description{
Get the contribution to the CRPS for different predictand classes.
}
\examples{
\dontrun{
data <- atmoswing::parseNcOutputs('optim/1/results', 1, 'validation')
crps.matrix <- atmoswing::crpsNbAnalogs(data)
res <- atmoswing::crpsContribClasses(crps.matrix$crps[, 30], data$target.values.raw)
barplot(res$labels)
title('Contribution to the CRPS (not cumulative)')
}

}
