% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crpsVectorWeightedByCriteria}
\alias{crpsVectorWeightedByCriteria}
\title{Process the CRPS weighted by the analogy criteria for the target period..}
\usage{
crpsVectorWeightedByCriteria(A, scale = 1, nb.analogs = 0, a = 0.44,
  b = 0.12)
}
\arguments{
\item{A}{Results of AtmoSwing as parsed by atmoswing::parseAllNcOutputs.}

\item{scale}{Importance given to the weights (0 to 1)}

\item{nb.analogs}{Number of analogs to consider (all of them if ignored or 0)}

\item{a, b}{Constants that have a law-dependent optimum from which the samples 
are derived.}
}
\value{
Vector of the CRPS value for every day of the target period.
}
\description{
Process the CRPS weighted by the analogy criteria for a given number of 
analogues for every day of the target period.
}
\examples{
\dontrun{
data <- atmoswing::parseAllNcOutputs('optim/1/results', 1, 'validation')
res <- atmoswing::crpsVectorWeightedByCriteria(data, 1, 30)
}

}
