% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{darkenColor}
\alias{darkenColor}
\title{Darken a given color.}
\usage{
darkenColor(color, factor = 1.4)
}
\arguments{
\item{color}{The color to darken}

\item{factor}{The factor used for darkening}
}
\value{
The darkened color
}
\description{
Darken a given color by a given factor.
}
\examples{
\dontrun{
color <- atmoswing::darkenColor("#AD71B5", 1.5)
}

}
