% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseMultipleRuns.R
\name{getAnnualTotals}
\alias{getAnnualTotals}
\title{Process the precipitation annual totals}
\usage{
getAnnualTotals(directory, datasets, methods, startYear, endYear)
}
\arguments{
\item{directory}{Root directory of multiple runs.}

\item{datasets}{List of datasets (must be used as folder names - e.g. /JRA-55/)}

\item{methods}{List of methods (must be used as folder names - e.g. /4Z/)}

\item{startYear}{First year of the total period.}

\item{endYear}{Last year of the total period.}
}
\value{
Dataframe with precip annual totals for every method / dataset.
}
\description{
Process the precipitation annual totals of multiple methods / datasets
}
\examples{
\dontrun{
datasets <- c('CFSR', 'ERA-20C', 'JRA-55')
methods <- c('2Z', '4Z', '4Z-2MI')
stations <- atmoswing::getAnnualTotals('path/to/runs', datasets, methods, 1981, 2010)
}

}
