% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseMultipleRuns.R
\name{getCrpss}
\alias{getCrpss}
\title{Process the CRPSS of multiple methods / datasets}
\usage{
getCrpss(directory, datasets, methods, period)
}
\arguments{
\item{directory}{Root directory of multiple runs.}

\item{datasets}{List of datasets (must be used as folder names - e.g. /JRA-55/)}

\item{methods}{List of methods (must be used as folder names - e.g. /4Z/)}

\item{period}{Either "calibration" or "validation".}
}
\value{
Dataframe with CRPSS scores for every method / dataset.
}
\description{
Process the CRPSS of multiple methods / datasets
}
\examples{
\dontrun{
datasets <- c('CFSR', 'ERA-20C', 'JRA-55')
methods <- c('2Z', '4Z', '4Z-2MI')
stations <- atmoswing::getCrpss('path/to/runs', datasets, methods, 'validation')
}

}
