% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseMultipleRuns.R
\name{getPercentageSimilarDates}
\alias{getPercentageSimilarDates}
\title{Get the percentage of similar analogue dates between reanalyses}
\usage{
getPercentageSimilarDates(directory, predictandDB, datasets, methods, period,
  q.threshold)
}
\arguments{
\item{directory}{Root directory of multiple runs.}

\item{predictandDB}{Path to the predictand DB.}

\item{datasets}{List of datasets (must be used as folder names - e.g. /JRA-55/)}

\item{methods}{List of methods (must be used as folder names - e.g. /4Z/)}

\item{period}{Either "calibration" or "validation".}

\item{q.threshold}{Optional precipitation threshold (for target dates). It is expressed as a quantile of days with precipitation (put -1 to ignore)}
}
\value{
Dataframe with the percentage of similar analogue dates
}
\description{
Get the percentage of similar analogue dates between reanalyses per method
}
\examples{
\dontrun{
datasets <- c('CFSR', 'ERA-20C', 'JRA-55')
methods <- c('2Z', '4Z', '4Z-2MI')
stations <- atmoswing::getBias('path/to/runs', 'path/to/db', datasets, methods, 'validation', 0.99)
}

}
