% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getReanalysesColors}
\alias{getReanalysesColors}
\title{Get the reference colors for multiple reanalyses.}
\usage{
getReanalysesColors(reanalyses)
}
\arguments{
\item{reanalyses}{A vector with the names of the reanalysis.}
}
\value{
A vector with the hex color codes
}
\description{
Get the reference colors for multiple reanalyses.
}
\examples{
\dontrun{
colors <- atmoswing::getReanalysesColors(c('CFSR', 'NR-1', 'JRA-55'))
}

}
