% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lightenColor}
\alias{lightenColor}
\title{Lighten a given color.}
\usage{
lightenColor(color, factor = 1.4)
}
\arguments{
\item{color}{The color to lighten}

\item{factor}{The factor used for lightening}
}
\value{
The lightened color
}
\description{
Lighten a given color by a given factor.
}
\examples{
\dontrun{
color <- atmoswing::lightenColor("#AD71B5", 1.5)
}

}
