% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalNbAnalogs.R
\name{optimalNbAnalog}
\alias{optimalNbAnalog}
\title{Identifies the optimal number of analogues on the crps matrix.}
\usage{
optimalNbAnalog(crps.matrix)
}
\arguments{
\item{crps.matrix}{Results of atmoswing::crpsNbAnalogs.}
}
\value{
Vector of the optimal number of analogues for each target date.
}
\description{
Identifies the optimal number of analogues on the crps matrix resulting from
atmoswing::crpsNbAnalogs.
}
\examples{
\dontrun{
data <- atmoswing::parseNcOutputs('optim/1/results', 1, 'validation')
res.crps <- atmoswing::crpsNbAnalogs(data)
res.nb <- atmoswing::optimalNbAnalog(res.crps$crps.smoothed)
}

}
