% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalNbAnalogs.R
\name{optimalNbAnalogOfAnalogs}
\alias{optimalNbAnalogOfAnalogs}
\title{Assign the optimal number of analogues to every analogue date}
\usage{
optimalNbAnalogOfAnalogs(A, target.a.nb)
}
\arguments{
\item{A}{Results of AtmoSwing as parsed by atmoswing::parseAllNcOutputs.}

\item{target.a.nb}{Results of atmoswing::optimalNbAnalog}
}
\value{
Matrix of the optimal number of analogues for each analogue date.
}
\description{
Assign the optimal number of analogues to every analogue date. This only
works on the calibration period, where the optimal number of analogues can
be assessed for the non independent dates.
}
\examples{
\dontrun{
data <- atmoswing::parseAllNcOutputs('optim/1/results', 1, 'validation')
res.crps <- atmoswing::crpsNbAnalogs(data)
target.a.nb <- atmoswing::optimalNbAnalog(res.crps$crps.smoothed)
res <- atmoswing::optimalNbAnalogOfAnalogs(data, target.a.nb$nb.min.med)
}

}
