% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutputs.R
\name{parseAllResultsText}
\alias{parseAllResultsText}
\title{Parse parameters resulting from AtmoSwing optimizer.}
\usage{
parseAllResultsText(directory, predictandDB, datasets, methods)
}
\arguments{
\item{directory}{Root directory of multiple runs.}

\item{predictandDB}{Path to the predictand DB.}

\item{datasets}{List of datasets (must be used as folder names - e.g. /JRA-55/)}

\item{methods}{List of methods (must be used as folder names - e.g. /4Z/)}
}
\value{
Results of the analogue method.
}
\description{
Extract resulting parameters from the text files resulting from AtmoSwing 
optimizer.
}
\examples{
\dontrun{
datasets <- c('CFSR', 'ERA-20C', 'JRA-55')
methods <- c('2Z', '4Z', '4Z-2MI')
data <- atmoswing::parseAllResultsText('path/to/runs', datasets, methods)
}

}
