% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutputs.R
\name{parseDatesNcFile}
\alias{parseDatesNcFile}
\title{Parse NetCDF files resulting from AtmoSwing optimizer.}
\usage{
parseDatesNcFile(filePath)
}
\arguments{
\item{filePath}{Path to the file containing the analog dates}
}
\value{
Results of the analogue method.
}
\description{
Extract results (for both analogues and the target situations: dates) from 
the NetCDF files resulting from AtmoSwing optimizer.
}
\examples{
\dontrun{
data <- atmoswing::parseDatesNcFile('path/to/results/validation/AnalogDates_id_291_step_0.nc')
}

}
