% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutputs.R
\name{parseScoresNcOutputs}
\alias{parseScoresNcOutputs}
\title{Parse NetCDF score files resulting from AtmoSwing optimizer.}
\usage{
parseScoresNcOutputs(directory, station.id, period, level = 1)
}
\arguments{
\item{directory}{Directory containing the outputs from AtmoSwing (containing 
the "calibration" or "validation" directories).}

\item{station.id}{ID of the station time series.}

\item{period}{Either "calibration" or "validation".}

\item{level}{Analogy level.}
}
\value{
Results of the analogue method.
}
\description{
Extract results (for the target situations: dates, predictand values, 
prediction score) from the NetCDF files resulting from AtmoSwing optimizer.
}
\examples{
\dontrun{
data <- atmoswing::parseScoresNcOutputs('optimizer-outputs/1/results', 1, 'validation')
}

}
