% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOutputs.R
\name{parseValuesNcFile}
\alias{parseValuesNcFile}
\title{Parse NetCDF files resulting from AtmoSwing optimizer.}
\usage{
parseValuesNcFile(filePath)
}
\arguments{
\item{filePath}{Path to the file containing the analog values}
}
\value{
Results of the analogue method.
}
\description{
Extract results (for both analogues and the target situations: predictand 
values) from the NetCDF files resulting from AtmoSwing optimizer.
}
\examples{
\dontrun{
data <- atmoswing::parseDatesNcFile('path/to/results/validation/AnalogValues_id_291_step_0.nc')
}

}
