% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{reanalysis.colors}
\alias{reanalysis.colors}
\title{Build a dataframe with the predefined color per reanalysis}
\format{An object of class \code{data.frame} with 10 rows and 2 columns.}
\usage{
reanalysis.colors
}
\value{
The colors dataframe.
}
\description{
Build a dataframe with the predefined color per reanalysis:
The S-RIP colour definitions for reanalysis datasets (Updated on 11 July 2017, for the S-RIP Full Report 2018)
https://s-rip.ees.hokudai.ac.jp/mediawiki/index.php/Notes_for_Authors
Reanalyses	R, G, B	Hexadecimal	Notes
MERRA-2	226, 31, 38	#E21F26	
MERRA	246, 153, 153	#F69999	
ERA-Interim	41, 95, 138	#295F8A	
ERA5	95, 152, 198	#5F98C6	
ERA-40	175, 203, 227	#AFCBE3	
JRA-55	114, 59, 122	#723B7A	
JRA-55C, JRA-55AMIP	173, 113, 181	#AD71B5	
JRA-25	214, 184, 218	#D6B8DA	
NCEP-NCAR R1	245, 126, 32	#F57E20	
NCEP-DOE R2	253, 191, 110	#FDBF6E	
20CR v2c	236, 0, 140	#EC008C	
20CR v2	247, 153, 209	#F799D1	
CERA-20C	0, 174, 239	#00AEEF	
ERA-20C	96, 200, 232	#60C8E8	
CFSR	52, 160, 72	#34A048	
REM	179, 91, 40	#B35B28	reanalysis ensemble mean
Other	255, 215, 0	#FFD700	
Observations	0, 0, 0	#000000	observations - black
Other observations	119, 119, 119	#777777	observations - grey
}
\keyword{datasets}
