% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{CRPS}
\alias{CRPS}
\title{Continuous ranked probability score}
\usage{
CRPS(x, x0, a = 0.44, b = 0.12)
}
\arguments{
\item{x}{vector of length n (number of members / analogues)}

\item{x0}{observed value}

\item{a}{constant that has a law-dependent optimum from which the samples are derived}

\item{b}{constant that has a law-dependent optimum from which the samples are derived}
}
\value{
Value of the CRPS score
}
\description{
Process the CRPS score for a given probability prediction compared to an observed value. 
The non-exceedence freequency is processed as follows: \code{"F(x) = (r-a)/(n+b)"}
}
\examples{
data <- read.table('testdaten.txt',header=F)
x <- sqrt(data$V5/70.82)
x0 <- sqrt(24.4/70.82)
result <- CRPS(x, x0, a=0.375, b=0.25)

}
