% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crpsNbAnalogs}
\alias{crpsNbAnalogs}
\title{Process the CRPS for an increasing number of analogues.}
\usage{
crpsNbAnalogs(A, filter.size = 9)
}
\arguments{
\item{A}{Results of AtmoSwing as parsed by atmoswingRToolbox::parseNcOutputs.}

\item{filter.size}{Length of "running window", has to be odd.}
}
\value{
Matrices with CRPS scores for an increasing number of analogues.
}
\description{
Process the CRPS for an increasing number of analogues for every day of the 
target period.
}
\examples{
data <- atmoswingRToolbox::parseNcOutputs('files/optimizer-outputs/1/results',
                                          1, 'validation')
res <- atmoswingRToolbox::crpsNbAnalogs(data, 9)

}
