% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crpsVector}
\alias{crpsVector}
\title{Process the CRPS for a given number of analogues.}
\usage{
crpsVector(A, nb.analogs = 0)
}
\arguments{
\item{A}{Results of AtmoSwing as parsed by atmoswing::parseNcOutputs.}

\item{nb.analogs}{Number of analogs to consider (all of them if ignored or 0)}
}
\value{
Matrices with CRPS scores for an increasing number of analogues.
}
\description{
Process the CRPS for a given number of analogues for every day of the 
target period.
}
\examples{
\dontrun{
data <- atmoswing::parseNcOutputs('optim/1/results', 1, 'validation')
res <- atmoswing::crpsVector(data, 30)
}

}
