% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceScore.R
\name{crpsVectorWeightedByCriteria}
\alias{crpsVectorWeightedByCriteria}
\title{Process the CRPS weighted by the analogy criteria for the target period..}
\usage{
crpsVectorWeightedByCriteria(A, nb.analogs = 0)
}
\arguments{
\item{A}{Results of AtmoSwing as parsed by atmoswing::parseNcOutputs.}

\item{nb.analogs}{Number of analogs to consider (all of them if ignored or 0)}
}
\value{
Vector of the CRPS value for every day of the target period.
}
\description{
Process the CRPS weighted by the analogy criteria for a given number of 
analogues for every day of the target period.
}
\examples{
\dontrun{
data <- atmoswing::parseNcOutputs('optim/1/results', 1, 'validation')
res <- atmoswing::crpsVector(data, 30)
}

}
