% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getReanalysisColor}
\alias{getReanalysisColor}
\title{Get the reference color for a given reanalysis.}
\usage{
getReanalysisColor(reanalysis)
}
\arguments{
\item{reanalysis}{The name of the reanalysis.}
}
\value{
The hex color code
}
\description{
Get the reference color for a given reanalysis.
}
\examples{
\dontrun{
col <- atmoswing::getReanalysisColor('CFSR')
}

}
